Created assemblies for each folder script

+++ Forgivest/Assets/Core/Scripts/Runtime/IDamage.cs
@@ -5,7 +5,7 @@ namespace Core
     public interface IDamage
     {
         bool IsCriticalHit { get; }
-        int Magnitude { get; }
+        float Magnitude { get; }
         GameObject Instigator { get; }
         object Source { get; }
     }
